making qelectrons a feature users can opt out of
