+++ covalent/executor/utils/wrappers.py
@@ -29,7 +29,7 @@ from typing import Any, Callable, Dict, List, Tuple
 
 import requests
 
-from covalent._shared_files.qelectron_utils import get_qelectron_db_path
+from covalent._shared_files.utils import get_qelectron_db_path
 from covalent._workflow.depsbash import DepsBash
 from covalent._workflow.depscall import RESERVED_RETVAL_KEY__FILES, DepsCall
 from covalent._workflow.depspip import DepsPip
