+++ requirements.txt
@@ -8,12 +8,8 @@ dask[distributed]>=2022.6.0
 fastapi>=0.100.0
 filelock>=3.12.2
 furl>=2.1.3
-lmdbm>=0.0.5
-mpire>=2.7.1
 natsort>=8.4.0
 networkx>=2.8.6
-orjson>=3.8.10
-pennylane>=0.31.1,<0.33.0
 psutil>=5.9.0
 pydantic>=2.1.1
 python-multipart>=0.0.6
