+++ setup.py
@@ -55,6 +55,12 @@ else:
 with open(requirements_file) as f:
     required = f.read().splitlines()
 
+# By default we install qelectron, but we can disable it by setting the environment variable
+disable_qelectron = os.environ.get("COVALENT_DISABLE_QELECTRON") == "1"
+if not disable_qelectron:
+    with open("requirements-qelectron.txt") as f:
+        required += f.read().splitlines()
+
 
 def recursively_append_files(directory: str):
     """
